Change invoices to always be vetted records.
Change invoice submission so that the record in Politeia is
automatically made vetted. This keeps all future changes to the
record in the same place, the vetted branch of the Politeia git repo.

Also fix a bug with converting the changes stream from the Politeia record
into an invoice.

Also ensure that server logging in the dataload util is persisted across
server restarts within the same execution.