Fix bug with route handling.
This commit fixes a bug with route handling where the request struct
for a given route was being reused across multiple requests to that same
route. For example, one request to /v1/invoices might pass a status,
and then the next request might not provide it. But the status would
still be in the request struct because it's being reused from the first
request.

This has been fixed by using the reflect package to construct an
instance of the request struct type when each request is made.