modules: Replace dep with Go's new modules solution
Go modules is an optional feature in Go 1.11 that is intended to play
the role that dep played.  This changes the project to use this new
method, a README update, and a fix to allow the project to work with the
latest changes to the goemail package in the same way that it once
worked.