Add ability to fetch dcrusd monthly rate.
This commit adds logic to pull DCR-BTC and BTC-USD price data from
the Binance and Kraken APIs (respectively) for 15 minute intervals
and store them in a file for the current month. It also supports
pulling data for the previous month if the current time is within the
first 5 days of the current month.

It also adds a route to fetch the DCR-USD rate, which reads all records
for the given month from the data file and calculates the rate by first
multiplying the average between the Open and Close values for a DCR-BTC
candlestick with the average between the Open and Close values for the
corresponding BTC-USD candlestick with the same timestamp. The logic
then averages DCR-USD values for all intervals in a given month and
returns it as the DCR-USD rate for the month.